Make the shims invokable as Node.js scripts
Some programs rely on the fact that the executable pointed to
by `which npm` is a valid JS file, as it is the case for normal
distributions of Node.js.

This patch turns the generated shims into Bash + Node.js > 0.12
polyglots, where the JS part just executes the file itself
(with `bash`, as specified by its shebang).