Change authorship
Sam wrote the original rbenv but not nodenv, so does the creator of the
upstream project get counted as author?

Will forked rbenv as nodenv so was therefor really the "creator" of
nodenv, but didn't originally author any of the code. Is Will the
author?

I'm the current maintainer and am the original publisher of nodenv on
npm. Am I the author?

DODGE THE QUESTION!

Remove author and just list all as contributors. :)