npm doesn't like symlinks
npm won't include symlinks in the tarball (it follows the symlinks and
includes the target files, not the links themselves). https://github.com/npm/npm/issues/3310

(seems fine, but also counter to the oft-repeated comment that "npm
isn't just for js but can also do plain files")

Regardless, since our symlink is only there for purposes of identifying
the binary, we can just manually declare that the nodenv bin points to
libexec/nodenv.