nodenv-version learns to describe aliases
When printing the current node version and its origin (how it was set),
nodenv-version now resolves any symlinks to determine the final
targeted version.

It prints the version name as selected, an arrow and the aliased version
name, and finally the origin.
If the version is not an alias, the output is the same as before.

This only affects the output of `nodenv version`;
`nodenv-version-name` and `nodenv-version-origin` are unaffected.