Ignore errors if nodenv-prefix fails
nodenv-prefix will print errors to STDERR if node isn't available in
PATH. (As demonstrated on Travis, since the nvm's nodes are removed from
PATH by the test-helper.)

We don't actually care if node is available in PATH or not for this
check, we only care to print the selected node and follow it if it's a
symlink. So if nodenv-prefix fails, the path would be empty, thus no
need to follow it and we just print the same message as before.