Make --version more semver friendly
The current output of nodenv --version is essentially the direct output
of git-describe, which is TAG-COMMITS_SINCE_TAG-gSHA.

The commits-since, and sha segments would be considered build metadata
in semver parlance.

This change tweaks the output of git-describe such that it is in a
semver compliant format; despite not strictly being semver compliant in
semantics. (Which isn't really possible with edge/HEAD anyway)

It also eliminates the "g" prefix which is used to denote the sha
references the git SCM.